feat(FTerm): Use FTerm INSTEAD for floating terminal. Because this would NOT "eat" the first character entered when the terminal buffer is respawned.
IMPORTANT NOTE:
The <leader>g option that could spawn "lazygit" is DISABLED inside an
opened terminal (a.k.a., "t" in keymap). This is done intentionally
since our <leader> is space and if we set this could make nvim wait for
our input every time we type in <leader>(space)!